feat: 🎸 merge docker and kubernetes provider
