+++ providers/container/src/main/resources/reference.conf
@@ -1,0 +1,30 @@
+unisonui {
+  providers += "tech.unisonui.providers.container.ContainerProvider"
+  provider.container {
+    kubernetes {
+      enabled = yes
+      polling-interval = "1 minute"
+    }
+
+    docker {
+      enabled = yes
+      host =  "unix:///var/run/docker.sock"
+    }
+
+    labels {
+      service-name = "unisonui.service-name"
+
+      openapi {
+        port  = "unisonui.specification.port"
+        protocol  = "unisonui.specification.protocol"
+        specification-path = "unisonui.specification.path"
+        use-proxy = "unisonui.specification.use-proxy"
+      }
+
+      grpc {
+        port = "unisonui.grpc.port"
+        tls = "unisonui.grpc.tls"
+      }
+    }
+  }
+}
