+++ providers/container/src/main/scala/tech/unisonui/providers/container/docker/client/models/Network.scala
@@ -1,0 +1,13 @@
+package tech.unisonui.providers.container.docker.client.models
+
+import io.circe.{Decoder, Encoder, HCursor, Json}
+
+final case class Network(ipAddress: String)
+
+object Network {
+  implicit val decoder: Decoder[Network] = (cursor: HCursor) =>
+    cursor.get[String]("IPAddress").map(Network(_))
+
+  implicit val encoder: Encoder[Network] = (network: Network) =>
+    Json.obj("IPAddress" -> Json.fromString(network.ipAddress))
+}
