+++ providers/container/src/main/scala/tech/unisonui/providers/container/docker/client/models/State.scala
@@ -1,0 +1,15 @@
+package tech.unisonui.providers.container.docker.client.models
+
+sealed trait State
+
+object State {
+  case object Start extends State
+  case object Stop  extends State
+
+  def fromString(string: String): Option[State] =
+    string match {
+      case "start" => Some(Start)
+      case "stop"  => Some(Stop)
+      case _       => None
+    }
+}
