+++ providers/container/src/main/scala/tech/unisonui/providers/container/settings/DockerSettings.scala
@@ -1,0 +1,3 @@
+package tech.unisonui.providers.container.settings
+
+final case class DockerSettings(host: String)
