+++ providers/container/src/main/scala/tech/unisonui/providers/container/settings/KubernetesSettings.scala
@@ -1,0 +1,4 @@
+package tech.unisonui.providers.container.settings
+
+import scala.concurrent.duration.FiniteDuration
+final case class KubernetesSettings(pollingInterval: FiniteDuration)
