--- providers/docker/src/main/resources/reference.conf
@@ -1,14 +1,0 @@
-unisonui {
-  providers += "tech.unisonui.providers.docker.DockerProvider"
-  provider.docker {
-    host =  "unix:///var/run/docker.sock"
-    labels {
-      port  = "unisonui.specification.port"
-      service-name = "unisonui.service-name"
-      specification-path = "unisonui.specification.path"
-      use-proxy = "unisonui.specification.use-proxy"
-      grpc-port = "unisonui.grpc.port"
-      grpc-tls = "unisonui.grpc.tls"
-    }
-  }
-}
