--- providers/kubernetes/src/main/resources/reference.conf
@@ -1,15 +1,0 @@
-unisonui {
-  providers += "tech.unisonui.providers.kubernetes.KubernetesProvider"
-  provider.kubernetes {
-    polling-interval = "1 minute"
-    labels {
-      service-name = "unisonui.service-name"
-      port  = "unisonui.specification.port"
-      protocol = "unisonui.specification.protocol"
-      specification-path = "unisonui.specification.path"
-      use-proxy = "unisonui.specification.use-proxy"
-      grpc-port = "unisonui.grpc.port"
-      grpc-tls = "unisonui.grpc.tls"
-    }
-  }
-}
